+++ app/src/main/java/com/six/tipsproject/arch/db/migration/User.kt
@@ -9,6 +9,6 @@ import android.arch.persistence.room.PrimaryKey
  * Created by Heavens on 2018-11-28.
  */
 @Entity(tableName = "user")
-data class User(@PrimaryKey(autoGenerate = true) @ColumnInfo(name = "uid") val uid: Long,
+data class User(@PrimaryKey(autoGenerate = true) @ColumnInfo(name = "uid") val uid: Int,
                 @ColumnInfo(name = "name") val name: String,
                 @ColumnInfo(name = "age") val age: Int = 0)
